+++ lib/startback/operation.rb
@@ -48,6 +48,10 @@ module Startback
       world.fetch(name){ super }
     end
 
+    def respond_to?(name)
+      super || (world && world.has_key?(name))
+    end
+
   protected
 
     def operation_world(op)